refactor: move code to src directory
